Inventory and Equipment
Preliminary User interfaces for the Inventory and Equipment screens are
mostly done- the Equipment screen doesn't display the item yet, but the
Inventory will display the item type. Click on an item in your inventory
to equip it to the equipment, click on an item in the equipment to
unequip it to the inventory. If an  equipment slot already has an item
of the type equipped, the currently equipped item will be replaced and
sent to the end of the inventory.

Pressing "I" opens the Inventory, pressing "U" opens the inventory. "1"
and "2" no longer do anything on press. I may end up typing the
inventory and equipment interfaces together so both are either open or
closed at any given time.