Inventory Interface Update
I imported the Necromancer GUI pack from the Asset Store and am using
some of the images and fonts that came with it for a nicer look to the
interfaces.

The inventory can now be clicked to come to the top layer and if clicked
and the mouse is dragged, will change their position as well. Camera
controls will be halted while the Inventory or Equipment menus are
active.

Most functionality is present, though a few visual tweaks and
potentially the addition of pages could help, but there are 30 inventory
spaces right now on the single page which should be pretty adequate for
the game I believe. I will probably add an Item information section to
the inventory when an item is clicked where you can see stats and stat
changes and options to equip or discard an item.

The test scene is set up with 31 Chests with a bunch of items for
testing the Inventory/Equipment. The Equipment Interface is still just
placeholder, but the buttons will unequip items still though it does not
display if one is equipped or not. An Equipment Interface Update will be
follow soonish.