Small Inventory Update
Added an item name and description to items- the name will be displayed
on buttons, and later on the description will be used when an item is
hovered over or selected. Also fixed an issue with unequipping items due
to Enum problems and fixed the background issue on the
Inventory/Equipment. The Chests have been updated to contain only one of
each item except amulets, which it has three different ones just to give
visual assurance that the appropriate item is actually being used.