Enemy Updates
Enemies will be using the NavMesh now, so there should be built-in
pathing and what not. Enemies will walk through obstacles to return to
their patrols right now since obstacle avoidance makes them take walk a
little strangely sometimes; I'll have to get some feedback on whether to
change that or not.

For some reason the Demon doesn't have upper body bones I could attach
its attack to, so it's using a broad trigger that turns on about when
the attack would hit the Player and covers the approximate area.

I'll look into smoothing out the transition from the Demon being in
attack range to the Player versus chasing him- right now if you run his
animations kind of stutter oddly, but my attempts to fix it ended up
having other issues so I'm just leaving it for now.