Two new enemies
I added in a flesh-golem type of enemy along with a ghoul; I've found
that using Vector3's for the waypoints is a bit annoying to edit, so
I'll be changing that in the next update to use Transforms. To avoid the
stuttering animation, we can use a delay equal to the attack's duration,
but would prefer a better solution.