Level 1 Implemented
Level one is filled in with an environment and enemies. The switches are
up and running, but patrol locations need work. May need some minor
rearrangements, but I don't think we have time. Added in the
Audio_Manager to play music depending on whether enemies are after you
or not. Added in a lot of assets, tried to reduce the number but there's
still a lot.